counts collections + resource backup

+++ registry-ws/src/main/java/org/gbif/registry/persistence/mapper/StaffMapper.java
@@ -29,12 +29,12 @@ public interface StaffMapper {
 
   List<Staff> search(@Nullable @Param("query") String query, @Nullable @Param("page") Pageable page);
 
-  int count();
+  long count();
 
-  int countWithFilter(@Nullable @Param("query") String query);
+  long countWithFilter(@Nullable @Param("query") String query);
 
-  int countByInstitution(@Param("institutionKey") UUID institutionKey);
+  long countByInstitution(@Param("institutionKey") UUID institutionKey);
 
-  int countByCollection(@Param("collectionKey") UUID collectionKey);
+  long countByCollection(@Param("collectionKey") UUID collectionKey);
 
 }
